+++ tests/middleware/test_logging.py
@@ -7,6 +7,15 @@ import websockets
 
 from tests.utils import run_server
 from uvicorn import Config
+from uvicorn.protocols.http.h11_impl import H11Protocol
+
+try:
+    from uvicorn.protocols.http.httptools_impl import HttpToolsProtocol
+except ImportError:  # pragma: nocover
+    HttpToolsProtocol = None
+
+
+HTTP_PROTOCOLS = [p for p in [H11Protocol, HttpToolsProtocol] if p is not None]
 
 
 @contextlib.contextmanager
@@ -27,7 +36,7 @@ async def app(scope, receive, send):
     await send({"type": "http.response.body", "body": b"", "more_body": False})
 
 
-@pytest.mark.asyncio
+@pytest.mark.anyio
 async def test_trace_logging(caplog, logging_config):
     config = Config(
         app=app, log_level="trace", log_config=logging_config, lifespan="auto"
@@ -48,8 +57,8 @@ async def test_trace_logging(caplog, logging_config):
         assert "ASGI [2] Completed" in messages.pop(0)
 
 
-@pytest.mark.asyncio
-@pytest.mark.parametrize("http_protocol", [("h11"), ("httptools")])
+@pytest.mark.anyio
+@pytest.mark.parametrize("http_protocol", HTTP_PROTOCOLS)
 async def test_trace_logging_on_http_protocol(http_protocol, caplog, logging_config):
     config = Config(
         app=app,
@@ -71,7 +80,7 @@ async def test_trace_logging_on_http_protocol(http_protocol, caplog, logging_con
         assert any(" - HTTP connection lost" in message for message in messages)
 
 
-@pytest.mark.asyncio
+@pytest.mark.anyio
 @pytest.mark.parametrize("ws_protocol", [("websockets"), ("wsproto")])
 async def test_trace_logging_on_ws_protocol(ws_protocol, caplog, logging_config):
     async def websocket_app(scope, receive, send):
@@ -107,7 +116,7 @@ async def test_trace_logging_on_ws_protocol(ws_protocol, caplog, logging_config)
         assert any(" - WebSocket connection lost" in message for message in messages)
 
 
-@pytest.mark.asyncio
+@pytest.mark.anyio
 @pytest.mark.parametrize("use_colors", [(True), (False), (None)])
 async def test_access_logging(use_colors, caplog, logging_config):
     config = Config(app=app, use_colors=use_colors, log_config=logging_config)
@@ -125,7 +134,7 @@ async def test_access_logging(use_colors, caplog, logging_config):
         assert '"GET / HTTP/1.1" 204' in messages.pop()
 
 
-@pytest.mark.asyncio
+@pytest.mark.anyio
 @pytest.mark.parametrize("use_colors", [(True), (False)])
 async def test_default_logging(use_colors, caplog, logging_config):
     config = Config(app=app, use_colors=use_colors, log_config=logging_config)
@@ -146,7 +155,7 @@ async def test_default_logging(use_colors, caplog, logging_config):
         assert "Shutting down" in messages.pop(0)
 
 
-@pytest.mark.asyncio
+@pytest.mark.anyio
 async def test_unknown_status_code(caplog):
     async def app(scope, receive, send):
         assert scope["type"] == "http"