+++ tests/middleware/test_wsgi.py
@@ -1,13 +1,15 @@
 import io
 import sys
-from typing import AsyncGenerator, List
+from typing import TYPE_CHECKING, AsyncGenerator, List
 
 import httpx
 import pytest
-from asgiref.typing import HTTPRequestEvent, HTTPScope
 
 from uvicorn._types import Environ, StartResponse
 from uvicorn.middleware.wsgi import WSGIMiddleware, build_environ
+
+if TYPE_CHECKING:
+    from asgiref.typing import HTTPRequestEvent, HTTPScope
 
 
 def hello_world(environ: Environ, start_response: StartResponse) -> List[bytes]:
@@ -50,7 +52,7 @@ def return_exc_info(environ: Environ, start_response: StartResponse) -> List[byt
         return [output]
 
 
-@pytest.mark.asyncio
+@pytest.mark.anyio
 async def test_wsgi_get() -> None:
     app = WSGIMiddleware(hello_world)
     async with httpx.AsyncClient(app=app, base_url="http://testserver") as client:
@@ -59,7 +61,7 @@ async def test_wsgi_get() -> None:
     assert response.text == "Hello World!\n"
 
 
-@pytest.mark.asyncio
+@pytest.mark.anyio
 async def test_wsgi_post() -> None:
     app = WSGIMiddleware(echo_body)
     async with httpx.AsyncClient(app=app, base_url="http://testserver") as client:
@@ -68,7 +70,7 @@ async def test_wsgi_post() -> None:
     assert response.text == '{"example": 123}'
 
 
-@pytest.mark.asyncio
+@pytest.mark.anyio
 async def test_wsgi_put_more_body() -> None:
     async def generate_body() -> AsyncGenerator[bytes, None]:
         for _ in range(1024):
@@ -81,7 +83,7 @@ async def test_wsgi_put_more_body() -> None:
     assert response.text == "123456789abcdef\n" * 64 * 1024
 
 
-@pytest.mark.asyncio
+@pytest.mark.anyio
 async def test_wsgi_exception() -> None:
     # Note that we're testing the WSGI app directly here.
     # The HTTP protocol implementations would catch this error and return 500.
@@ -91,7 +93,7 @@ async def test_wsgi_exception() -> None:
             await client.get("/")
 
 
-@pytest.mark.asyncio
+@pytest.mark.anyio
 async def test_wsgi_exc_info() -> None:
     # Note that we're testing the WSGI app directly here.
     # The HTTP protocol implementations would catch this error and return 500.
@@ -114,7 +116,7 @@ async def test_wsgi_exc_info() -> None:
 
 
 def test_build_environ_encoding() -> None:
-    scope: HTTPScope = {
+    scope: "HTTPScope" = {
         "asgi": {"version": "3.0", "spec_version": "2.0"},
         "scheme": "http",
         "raw_path": b"/\xe6\x96\x87",
@@ -129,7 +131,7 @@ def test_build_environ_encoding() -> None:
         "headers": [(b"key", b"value1"), (b"key", b"value2")],
         "extensions": {},
     }
-    message: HTTPRequestEvent = {
+    message: "HTTPRequestEvent" = {
         "type": "http.request",
         "body": b"",
         "more_body": False,