+++ tests/test_cli.py
@@ -122,6 +122,21 @@ def test_cli_incomplete_app_parameter() -> None:
     assert result.exit_code == 1
 
 
+def test_cli_event_size() -> None:
+    runner = CliRunner()
+
+    with mock.patch.object(main, "run") as mock_run:
+        result = runner.invoke(
+            cli,
+            ["tests.test_cli:App", "--h11-max-incomplete-event-size", str(32 * 1024)],
+        )
+
+    assert result.output == ""
+    assert result.exit_code == 0
+    mock_run.assert_called_once()
+    assert mock_run.call_args[1]["h11_max_incomplete_event_size"] == 32768
+
+
 @pytest.fixture()
 def load_env_h11_protocol():
     old_environ = dict(os.environ)